+++ test-compute-kernels/src/cacheop_flev_1mps_loop_shires/cacheop_flev_1mps_loop_shires.c
@@ -132,7 +132,7 @@ uint64_t generate_random_value(uint64_t lfsr)
     register const uint64_t polynomial = POLYNOMIAL_BIT;
 
 // Minion are slow to branch so unroll this loop
-#pragma GCC unroll 35
+PRAGMA_UNROLL_LOOP(35)
     for (int i = 0; i < LFSR_SHIFTS_PER_READ; i++)
     {
 #ifdef ASM_LFSR