+++ test-compute-kernels/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Fixed
 - [SW-19533] Add support to the kernels to be clang compatible
+- [SW-19783] Added clang loop unroll support
 ### Security
 
 ## [1.10.0] - 2023-12-07